refactor(api): renaming candidate repository
