+++ api/src/certification/evaluation/domain/services/index.js
@@ -11,8 +11,8 @@ import * as competenceMarkRepository from '../../../shared/infrastructure/reposi
 import * as complementaryCertificationBadgesRepository from '../../../shared/infrastructure/repositories/complementary-certification-badge-repository.js';
 import * as complementaryCertificationCourseResultRepository from '../../../shared/infrastructure/repositories/complementary-certification-course-result-repository.js';
 import * as calibratedChallengeRepository from '../../infrastructure/repositories/calibrated-challenge-repository.js';
+import * as candidateRepository from '../../infrastructure/repositories/candidate-repository.js';
 import * as certificationAssessmentHistoryRepository from '../../infrastructure/repositories/certification-assessment-history-repository.js';
-import * as certificationCandidateRepository from '../../infrastructure/repositories/certification-candidate-repository.js';
 import * as challengeCalibrationRepository from '../../infrastructure/repositories/challenge-calibration-repository.js';
 import * as complementaryCertificationScoringCriteriaRepository from '../../infrastructure/repositories/complementary-certification-scoring-criteria-repository.js';
 import * as scoringConfigurationRepository from '../../infrastructure/repositories/scoring-configuration-repository.js';
@@ -39,7 +39,7 @@ import { handleV3CertificationScoring } from './scoring/scoring-v3.js';
  * @typedef {scoringService} ScoringService
  * @typedef {areaRepository} AreaRepository
  * @typedef {placementProfileService} PlacementProfileService
- * @typedef {certificationCandidateRepository} CertificationCandidateRepository
+ * @typedef {candidateRepository} CandidateRepository
  * @typedef {complementaryCertificationBadgesRepository} ComplementaryCertificationBadgesRepository
  * @typedef {certificationAssessmentRepository} CertificationAssessmentRepository
  * @typedef {complementaryCertificationCourseResultRepository} ComplementaryCertificationCourseResultRepository
@@ -61,7 +61,7 @@ const dependencies = {
   areaRepository,
   placementProfileService,
   scoringService,
-  certificationCandidateRepository,
+  candidateRepository,
   complementaryCertificationBadgesRepository,
   certificationAssessmentRepository,
   complementaryCertificationCourseResultRepository,
